backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
openai>=1.40.0
anthropic>=0.40.0
google-genai>=1.2.0
python-dotenv>=1.0.0
httpx>=0.27.0
pytest>=8.0.0
# CI(ubuntu, Python 3.11)와 최신 파이썬(3.13+) 모두에서 휠이 있는 버전대로 해석되도록 하한만 둔다.
torch>=2.5.1
scikit-learn>=1.3.0
# 선택: TEAM_SEMANTIC_ENCODER=1 시 자기서술 문장 임베딩(다국어 MiniLM 등) — 기본 0은 미사용
sentence-transformers>=3.0.0
